requirements.txt
pandas
numpy
MetaTrader5
openpyxl
python-dotenv
requests
